Started refactoring handlers
