+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/handler/BackgroundTaskHandler.java
@@ -1,0 +1,36 @@
+package edu.byu.cs.tweeter.client.model.service.handler;
+
+import android.os.Bundle;
+import android.os.Handler;
+import android.os.Message;
+
+import androidx.annotation.NonNull;
+
+import edu.byu.cs.tweeter.client.model.service.backgroundTask.BackgroundTask;
+import edu.byu.cs.tweeter.client.model.service.observer.ServiceObserver;
+
+public abstract class BackgroundTaskHandler<T extends ServiceObserver> extends Handler {
+
+    private T observer;
+
+    public BackgroundTaskHandler(T observer) {
+        this.observer = observer;
+    }
+
+    @Override
+    public void handleMessage(@NonNull Message msg) {
+        boolean success = msg.getData().getBoolean(BackgroundTask.SUCCESS_KEY);
+        if (success) {
+            handleSuccess(msg.getData(), observer);
+//            observer.handleSuccess();
+        } else if (msg.getData().containsKey(BackgroundTask.MESSAGE_KEY)) {
+            String message = msg.getData().getString(BackgroundTask.MESSAGE_KEY);
+            observer.handleFailure(message);
+        } else if (msg.getData().containsKey(BackgroundTask.EXCEPTION_KEY)) {
+            Exception ex = (Exception) msg.getData().getSerializable(BackgroundTask.EXCEPTION_KEY);
+            observer.handleException(ex);
+        }
+    }
+
+    protected abstract void handleSuccess(Bundle data, T observer);
+}
