+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/handler/SimpleNotificationHandler.java
@@ -1,0 +1,17 @@
+package edu.byu.cs.tweeter.client.model.service.handler;
+
+import android.os.Bundle;
+
+import edu.byu.cs.tweeter.client.model.service.observer.SimpleNotificationObserver;
+
+public class SimpleNotificationHandler extends BackgroundTaskHandler<SimpleNotificationObserver>{
+
+    public SimpleNotificationHandler(SimpleNotificationObserver observer) {
+        super(observer);
+    }
+
+    @Override
+    protected void handleSuccess(Bundle data, SimpleNotificationObserver observer) {
+        observer.handleSuccess();
+    }
+}
