+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/observer/ServiceObserver.java
@@ -1,0 +1,9 @@
+package edu.byu.cs.tweeter.client.model.service.observer;
+
+import java.security.spec.ECField;
+
+public interface ServiceObserver {
+
+    void handleFailure(String message);
+    void handleException(Exception exception);
+}
