+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/observer/SimpleNotificationObserver.java
@@ -1,0 +1,5 @@
+package edu.byu.cs.tweeter.client.model.service.observer;
+
+public interface SimpleNotificationObserver extends ServiceObserver{
+    void handleSuccess();
+}
